// src/pages/modules/SIM-RS/pharmacy/medicine/data/types.ts
export interface IMedicine {
  id_obat: string
  id_satuan_organisasi: string
  nama_obat: string
  kategori_obat: string
  bentuk_sediaan: string
  satuan: string
  harga: number
  deskripsi: string
  is_status: boolean
  created_at: string
  created_user: string
  updated_at: string
  updated_user: string
  nama_satuan_organisasi: string
  nama_user_created: string
  nama_user_updated: string
}